Added game in hand stats and win shares
Created functions in statfunctions to calculate game in hand stats (GIHWR on 17lands) and win shares. Made the CardDerivedStats table in tabledefinitions to hold these stats and a couple other planned stats. Currently have the ability to load the GIH stats in to that table. Still need to load in win shares and create access functions.