Changes to handle MKM
Changes, mainly to setinfo, to be able to handle the different types of cards that are present in MKM such as split cards and list cards. MKM data should now be up on the server.
cardinfopatch is a file to manually input the cards that are not found on scryfall.